requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
setuptools>=65.0.0,<74
langchain>=0.1.0
langchain-google-genai>=2.0.0
langchain-chroma>=0.1.0
tensorflow>=2.15.0
tensorflow-hub>=0.15.0
chromadb>=0.4.22
